Refactoring into `plugins.module_utils.pool` has been reverted
