--- plugins/module_utils/entry.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import, division, print_function
-__metaclass__ = type
-
-
-class EntryNotFound(Exception):
-    pass
